requirements.txt
requests==2.27.1
quart==0.18.0
quart-auth==0.7.0
python-socketio==5.5.2
protobuf==4.21.6